need int64

+++ statsdaemon.go
@@ -601,9 +601,9 @@ func parseLine(line []byte) *Packet {
 	}
 }
 
-func round(val float64) int {
-    if val < 0 { return int(val-0.5) }
-    return int(val+0.5)
+func round(val float64) int64 {
+    if val < 0 { return int64(val-0.5) }
+    return int64(val+0.5)
 }
 
 func logParseFail(line []byte) {
